requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg==3.1.16
psycopg-binary==3.1.10
psycopg2-binary==2.9.9
python-dotenv==1.0.0
regex==2023.5.5
SQLAlchemy==2.0.24
typing_extensions==4.6.2
tzdata==2023.3
Werkzeug==2.3.4
